+++ src/Template/Renderer.php
@@ -13,39 +13,90 @@ namespace Rougin\Slytherin\Template;
 class Renderer implements RendererInterface
 {
     /**
-     * @var array
+     * @var string[]
      */
     protected $paths = array();
 
     /**
-     * Initializes the directory instance.
+     * Initializes the renderer instance.
      *
      * @param array|string $paths
      */
     public function __construct($paths)
     {
-        $paths = (array) $paths;
-
-        $this->paths = $paths;
+        $this->paths = (array) $paths;
     }
 
     /**
-     * Renders a template.
+     * Renders a file from a specified template.
      *
      * @param  string $template
      * @param  array  $data
      * @return string
+     *
+     * @throws \InvalidArgumentException
      */
     public function render($template, array $data = array())
     {
-        $file = $this->find($template . '.php');
+        $name = (string) str_replace('.', '/', $template);
 
-        // Extracts the specific parameters to the template.
+        foreach ((array) $this->paths as $key => $path) {
+            $files = (array) $this->files($path);
+
+            $item = $this->check($files, $path, $key, $name . '.php');
+
+            $item !== null && $file = $item;
+        }
+
+        if (isset($file) === false) {
+            $message = 'Template file "%s" not found.';
+
+            $message = sprintf($message, $name);
+
+            throw new \InvalidArgumentException($message);
+        }
+
+        return $this->extract($file, $data);
+    }
+
+    /**
+     * Checks if the specified file exists.
+     *
+     * @param  array          $files
+     * @param  string         $path
+     * @param  string|integer $source
+     * @param  string         $template
+     * @return string|null
+     */
+    protected function check(array $files, $path, $source, $template)
+    {
+        foreach ((array) $files as $key => $value) {
+            $filepath = str_replace($path, $source, $value);
+
+            $filepath = str_replace('\\', '/', (string) $filepath);
+
+            $filepath = preg_replace('/^\d\//i', '', $filepath);
+
+            strtolower($filepath) === $template && $file = $value;
+        }
+
+        return isset($file) === true ? $file : null;
+    }
+
+    /**
+     * Extracts the contents of the specified file.
+     *
+     * @param  string $filepath
+     * @param  array  $data
+     * @return string
+     */
+    protected function extract($filepath, array $data)
+    {
         extract($data);
 
         ob_start();
-        
-        include $file;
+
+        include $filepath;
 
         $contents = ob_get_contents();
 
@@ -55,28 +106,19 @@ class Renderer implements RendererInterface
     }
 
     /**
-     * Finds the specified template from the list of paths.
+     * Returns an array of filepaths from a specified directory.
      *
-     * @param  string      $template
-     * @param  string|null $file
-     * @return string
-     *
-     * @throws \InvalidArgumentException
+     * @param  string $path
+     * @return string[]
      */
-    protected function find($template, $file = null)
+    protected function files($path)
     {
-        foreach ((array) $this->paths as $path) {
-            $files = glob($path . '/' . $template);
+        $directory = new \RecursiveDirectoryIterator($path);
 
-            empty($files) || $file = $files[0];
-        }
+        $iterator = new \RecursiveIteratorIterator($directory);
 
-        if (is_null($file) === true) {
-            $message = 'Template file not found.';
+        $regex = new \RegexIterator($iterator, '/^.+\.php$/i', 1);
 
-            throw new \InvalidArgumentException($message);
-        }
-
-        return $file;
+        return (array) array_keys(iterator_to_array($regex));
     }
 }